[skip ci] add d3.t as test script
shows two ways to use IPC::Run to capture the output of running the script.
When there's a simple makefile with `make test` that runs `prove -vl d3.t`, `cover -test` will properly run the test and the report will include the spawned processes.  (It actually includes the prove.bat and d3.t in the coverage report when run that way, which is slightly different than Makefile.PL->Makefile->`gmake testcover`=`cover -test` ... so maybe the way the Makefile.PL-based Makefile phrases the `test:` target is different.  I _believe_ that a generated Makefile will still have the coverage report for all the spawned code, not just the libraries covered by the .t files... but I might want to check that when I have a chance.)